+++ src/holyc_lib.mli
@@ -114,6 +114,7 @@ end
 module Ir_integer_interpreter = Ir.Integer_interpreter
 module Ir_runtime_call_context = Ir.Runtime_call_context
 module Native_parameter_defaults = Driver.Native_parameter_defaults
+module Native_global_initializers = Driver.Native_global_initializers
 module Ir_integer_program_lowering = Ir.Integer_program_lowering
 module Ir_global_initialization = Ir.Global_initialization